refactor(observability): split assembly.rs tests back under the ceiling
The attachment and summary-serialization records took assembly.rs to 606.
Its last two tests -- binding-summary merge order and invalid-summary
rejection -- move to assembly/tests/binding_summaries.rs, following the
nested test-module convention this file already uses for
assembly/tests/workspace_attachment.rs. 606 -> 553.

Two other files still fail this check on this PR:
delivery/runtime/tests.rs (656) and TranscriptToolCallItemBlock.test.tsx
(609). Both fail identically on the base branch with this PR removed, so
they belong to #1776, not here.
